refactor: change imports because file name changes in both src code and test code
