+++ src/app.ts
@@ -2,7 +2,6 @@ import express, { Request, Response, NextFunction } from 'express';
 import { errorHandler } from './utils/errors/error-handler.middleware';
 import cors from "cors";
 import postRoutes from './features/posts/routes/post.routes'; // Adjust the path if your routes file is elsewhere
-import userPostsRouter from './features/posts/routes/userPosts.routes';
 
 export const app = express();
 const PORT = 3000;
@@ -13,8 +12,6 @@ app.get('/', (req, res) => {
 app.use(express.json());
 app.use(cors());
 app.use('/api', postRoutes);
-// Add the user posts routes
-app.use('/api', userPostsRouter);
 
 // Error handling middleware should be last
 app.use((err: Error, req: Request, res: Response, next: NextFunction) => {
